+++ docs/changelog.rst
@@ -36,6 +36,28 @@ modules like ``requests`` (issue #19).
 This feature may be resurrected in a later version if a safe implementation can be found.
 
 
+Deprecated ``isinstance`` removed
+---------------------------------
+
+``future`` v0.8.2 briefly introduced a replacement for the ``isinstance``
+builtin. This was then removed and its use was deprecated as of v0.9.0.
+The alias for the builtin ``isinstance`` has now been removed from
+``future.builtins``.
+
+
+Internal changes
+----------------
+
+The internal ``future.builtins.backports`` module has been renamed to
+``future.builtins.types``. This will change the ``repr`` of ``future``
+types but not their use.
+
+
+.. whats-new-0.10.2:
+
+What's new in version 0.10.2
+============================
+
 New context manager for standard_library hooks
 ----------------------------------------------